+++ packages/app/src/hooks/useAverageRewardRate/index.ts
