+++ packages/app/src/hooks/useAverageRewardRate/types.ts
