+++ packages/app/src/overlay/modals/RewardCalculator/index.tsx
@@ -7,7 +7,7 @@ import { getChainIcons } from 'assets'
 import { getStakingChainData } from 'consts/util'
 import { useNetwork } from 'contexts/Network'
 import { useValidators } from 'contexts/Validators/ValidatorEntries'
-import { useAverageRewardRate } from 'hooks/useAvgRewardRate'
+import { useAverageRewardRate } from 'hooks/useAverageRewardRate'
 import { Balance } from 'library/Balance'
 import { Title } from 'library/Modal/Title'
 import type { ChangeEvent } from 'react'
